feat: modulize component done
